Lab2 improved saving algoritgm
Now timers continue after loaded.

+++ Lab2/Timers/alarmclock.h
@@ -37,7 +37,7 @@ public:
      */
     bool repeated() override;
 
-    int asNum() override;
+    qint64 asNum() override;
 
 protected slots:
     /**
